Separate apiserver handler filters
