+++ pkg/apiserver/filters/impersonation.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apiserver
+package filters
 
 import (
 	"fmt"
